requirements.txt
langchain==0.3.0
langchain-community==0.3.0
langchain-google-genai==2.0.0
sentence-transformers==3.0.1
chromadb==0.5.5
pypdf==5.0.0
gradio==4.44.0
python-dotenv==1.0.1